@@ -3,11 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@blueprintjs/core": "^3.52.0",
-    "@blueprintjs/popover2": "^0.12.9",
-    "@blueprintjs/select": "^3.18.11",
-    "@chakra-ui/layout": "^1.7.4",
-    "@chakra-ui/system": "^1.10.3",
+    "@blueprintjs/core": "^4.3.2",
+    "@blueprintjs/popover2": "^1.2.2",
+    "@blueprintjs/select": "^4.2.2",
+    "@chakra-ui/layout": "^2.0.0",
+    "@chakra-ui/system": "^2.0.2",
     "@dnd-kit/core": "^5.0.1",
     "@dnd-kit/modifiers": "^5.0.0",
     "@dnd-kit/sortable": "^6.0.0",
@@ -20,31 +20,18 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.15",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
     "@vitejs/plugin-react": "^1.3.2",
+    "electron-context-menu": "^3.1.2",
     "inobounce": "^0.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "rxjs": "^7.5.4",
     "sass": "^1.49.7",
     "typescript": "^4.5.5",
     "vite": "^2.9.9",
-    "vite-plugin-env-compatible": "^1.1.1",
-    "web-vitals": "^2.1.4",
-    "workbox-background-sync": "^6.4.2",
-    "workbox-broadcast-update": "^6.4.2",
-    "workbox-cacheable-response": "^6.4.2",
-    "workbox-core": "^6.4.2",
-    "workbox-expiration": "^6.4.2",
-    "workbox-google-analytics": "^6.4.2",
-    "workbox-navigation-preload": "^6.4.2",
-    "workbox-precaching": "^6.4.2",
-    "workbox-range-requests": "^6.4.2",
-    "workbox-routing": "^6.4.2",
-    "workbox-strategies": "^6.4.2",
-    "workbox-streams": "^6.4.2"
+    "vite-plugin-env-compatible": "^1.1.1"
   },
   "scripts": {
     "start": "vite",
@@ -67,5 +54,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "electron": "^18.2.4",
+    "vite-plugin-electron": "^0.4.4",
+    "vite-plugin-pwa": "^0.12.0"
   }
 }
