@@ -1,5 +1,6 @@
 import { Subject } from "rxjs";
 import { DrakModePreference } from "./darkmode";
+import { ILogRecord } from "./ILogFrame";
 import { RxAtom } from "./RxAtom";
 
 function storedGlobalState<T>(key: string, defaultValue?: T): RxAtom<T> {
@@ -26,10 +27,10 @@ export const currentTab = new RxAtom<CurrentTab>(null);
 export const updateAvailiable = new RxAtom<boolean>(false);
 export const currentDevice = new RxAtom<string>(null);
 export const dispatcherState = new RxAtom<string>('running');
-export const colorScheme = storedGlobalState<DrakModePreference>("darkmode", "system");
+export const colorSchemePreference = storedGlobalState<DrakModePreference>("darkmode", "system");
 
 export const logScrollbackLimit = storedGlobalState<number>("logScrollbackLimit", 1500);
 export const showAboutOnStartup = storedGlobalState<boolean>("showAboutOnStartup", true);
 export const autoUpdate = storedGlobalState<boolean>("autoUpdate", true);
 
-export const log$ = new Subject();
+export const log$ = new Subject<ILogRecord>();
