@@ -110,7 +110,7 @@ const TaskItem = React.forwardRef((props: ITaskItemProps, ref) => {
 
   const textElm = useMemo(() => {
     const muted = itemStatus === TaskItemStatus.Skip || itemStatus === TaskItemStatus.Stop;
-    return <Text ellipsize={true} className={"task-item-title flex-grow-1" + (muted ? ' bp3-text-muted' : '')}>{title}</Text>
+    return <Text ellipsize={true} className={"task-item-title flex-grow-1" + (muted ? ' bp4-text-muted' : '')}>{title}</Text>
   }, [itemStatus, title]);
 
   let classes = ['task-item'];
@@ -220,7 +220,7 @@ const TaskList: React.FunctionComponent<ITaskListProps> = function (props) {
 
   return (
     <DnD.DndContext sensors={sensors} collisionDetection={DnD.closestCenter} modifiers={[restrictToVerticalAxis]} onDragEnd={handleDragEnd} onDragStart={handleDragStart}>
-      <Card style={{ ...props.style, padding: "0", flexGrow: 1 }} className='tasklist-widget default-background' >
+      <Card style={{ ...props.style, flexGrow: 1 }} className='tasklist-widget default-background card-no-padding' >
         <Flex position="relative" w="100%" h="100%" flexDirection="column" className='scroller overflow-y-scroll'>
           <DnDSortable.SortableContext items={props.tasks} strategy={DnDSortable.verticalListSortingStrategy}  >
 
